fix(install): avoid duplicate results when fetching recommendations
Without this change, the duplicate fix in 9ee4a93 is too aggressive in removing
the duplicates.  In the case where we fetch recipes, we want all results so that
we are able to effectively determine all of the process filters regardless of
the name of the recipe.  Here we shift that duplicate avoidance to the
FetchRecommendations call instead, since we only want to remove the duplicates
in the case the recipe service sends back a duplicate when we are looking for
recommendations.