chore(install): implement advanced mode
Here we implement a new flag and adjust the default workflow when the flag is
not present.  When the "advanced" flag is used, we prefer to prompt for
everything, with the exception of installing the infra agent, which will always
be installed.  Without this flag enabled, the default flow now includes fewer
prompts, only asking for variables that are not available from the environment,
and assuming that all OHI that are discovered available should be installed.