Update install.sh to use optional proxy
Added detection of the HTTPS_PROXY env variable so that the curl flags can be set to use the proxy if it was specified. The guided installation for the Infra agent will need to be updated afterwards to set the environment variable in the command line for the initial curl command and the bash shell that it gets piped into.